Better matching for jdbc provider/datasource
This commit will match for existing datasources and providers that
aren't necessarily wrapped in double quotes.  wsadmin seems to return
the names of these wrapped in quotes if they include spaces.  If they
don't, they aren't quoted.  This fixes the regex used to match the
existence to make the double quote optional.