(FM-5980) Find wsadmin script in profile w/backup
When collecting cluster member resources, the resources set the
`profile` paramater to the AppServer profile, which is not installed
on the DMGR agent. The websphere_helper was looking at the `profile`
resource parameter with a fallback to `dmgr_profile` resource
parameter when `profile` was not set. This caused the wsadmin.sh
script to not be found as the path was incorrect.

Instead search for wsadmin.sh in a path that contains the `profile`
resource parameter when it is available first, but fall back to
`dmgr_profile` resource paramater for the path when the first path
does not exist or `profile` is nil.

If after setting path with a fallback the file still does not exist,
raise a Puppet error indicating as much with the paths that were
checked.