(maint) Update Rakefile for modsync 915cde70e20
As the Rakefile is not managed by modsync, when the changes for commit 915...
were applied, the Rakefile was incorrect.  This commit adds a guard for
blacksmith tasks so they do to load and cause errors if the blacksmith gem is
not bundled.  This is the case for with --without system_tests is specified.