(maint) Update Rakefile for metadata-lint gem
Previously the rakefile would fail to compile any task if the metadata-lint
gem did not have its tasks loaded or the gem was missing.  This commit adds a
guard to the step which clears any previous defintion to only do this if the
task already exists.  Otherwise Rake would throw a Runtime Error.